+++ packages/dashboard/chart/scatter-chart/easypack.config.ts
@@ -2,9 +2,18 @@
  * @easy-editor/easypack configuration
  * @type {import('@easy-editor/easypack').EasypackConfig}
  */
+import { ECHARTS_EXTERNALS, ECHARTS_GLOBALS, echartsExternalPlugin } from './.vite/echarts-external-plugin'
+
 export default {
   preset: 'material',
   dev: {
     port: 5001,
   },
+  rollup: {
+    external: ECHARTS_EXTERNALS,
+    output: {
+      globals: ECHARTS_GLOBALS,
+    },
+  },
+  vitePlugins: [echartsExternalPlugin()],
 }